Update the logo size of the login selection screen when open on tablet device
