+++ app/components/shared/GradientScrollViewComponent.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import {ScrollView, StyleSheet, Text} from 'react-native';
+import {ScrollView, StyleSheet} from 'react-native';
 import LinearGradient from 'react-native-linear-gradient';
 import color from '../../themes/color';
 import {screenHorizontalPadding, scrollViewPaddingBottom} from '../../constants/component_constant';
 
-const GradientBackgroundComponent = (props) => {
+const GradientScrollViewComponent = (props) => {
   return (
     <LinearGradient
       colors={[color.primaryColor, 'rgba(170, 73, 133, 0.88)']}
@@ -30,4 +30,4 @@ const styles = StyleSheet.create({
   }
 });
 
-export default GradientBackgroundComponent;
+export default GradientScrollViewComponent;
